requirements.txt
django-extensions==3.2.1
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-jsonapi==6.0.0
inflection==0.5.1
pytz==2022.6
sqlparse==0.4.3
tzdata==2022.6